% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis.R
\name{BayesAnalysis}
\alias{BayesAnalysis}
\title{BayesAnalysis: Bayes analysis utilinsing MCMCpack}
\usage{
BayesAnalysis(model, data, choice = "beta", iter = 10000, ...)
}
\arguments{
\item{model}{Model to test}

\item{data}{data for model}

\item{choice}{choice on how to choose b0}

\item{iter}{iterations of MCMC}

\item{...}{other parameters to use in MCMCregress()}
}
\value{
Fancy list
}
\description{
BayesAnalysis: Bayes analysis utilinsing MCMCpack
}
\examples{
## Example 1
BayesAnalysis(QUALITY ~ TEMP + PRESSURE, PRODQUAL, choice = "mean", iter = 10000)
}
