% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BROWSER.R
\docType{data}
\name{BROWSER}
\alias{BROWSER}
\title{Browser Data information}
\format{
A data frame with 16 rows and 4 column(s)
\describe{
  \item{TASK}{Number of each test}
  \item{TIME}{Time for browser to complete a task, in hundredths of seconds}
  \item{NUMBER}{Number of Objects}
  \item{BROWSER}{Browser used in task}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
BROWSER
}
\description{
Data set of Browser test information.
   Data frame of 16 rows and 4 columns
}
\keyword{datasets}
