% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUTTERFAT.R
\docType{data}
\name{BUTTERFAT}
\alias{BUTTERFAT}
\title{BUTTERFAT}
\format{
An object of class \code{data.frame} with 100 rows and 3 columns.
}
\source{
s20x Package
}
\usage{
BUTTERFAT
}
\description{
BUTTERFAT
}
\details{
This is a copy of `butterfat.df` from the s20x Package

\describe{
  \item{}{See: \link[s20x]{butterfat.df}}
}
}
\keyword{datasets}
