% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRIME.R
\docType{data}
\name{CRIME}
\alias{CRIME}
\title{2019 Crime Data information}
\format{
A data frame with 52 rows and 12 column(s)
\describe{
  \item{Area}{State or Teritory}
  \item{Year}{Year Crime data was taken}
  \item{Population}{Population of a given area}
  \item{Violent_Crime}{Total Violent Crime per 100000 in an area}
  \item{Murder_and_manslaughter}{Murder and Manslaughter cases per 100000 in an area}
  \item{Rape}{Rape cases per 100000 in an area}
  \item{Robbery}{Robbery cases per 100000 in an area}
  \item{Aggravated_Assault}{Aggriavated Assault cases per 100000 in an area}
  \item{Property_Crime}{Property Crime cases per 100000 in an area}
  \item{Burglary}{Burglary Cases per 100000 in an area}
  \item{Larceny_Theft}{Larceny Theft cases per 100000 in an area}
  \item{Motor}{Motor vehicle per 100000 in an area}

}
}
\source{
\url{https://ucr.fbi.gov/crime-in-the-u.s/2019/crime-in-the-u.s.-2019/topic-pages/violent-crime}
}
\usage{
CRIME
}
\description{
Data set of Crime in the United States
   Data frame of 52 rows and 12 columns
}
\keyword{datasets}
