% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRBootstrap.R
\name{MLRBootstrap}
\alias{MLRBootstrap}
\title{MLRBootstrap: Bootstrap analysis of MLR model}
\usage{
MLRBootstrap(model, data, iter = 5000, alpha = 0.05, HistBreaks = 30)
}
\arguments{
\item{model}{Model to do bootstrap analysis}

\item{data}{Data for bootstrap analysis}

\item{iter}{Iterations of the bootstrap}

\item{alpha}{Used to create 1-alpha CI and etc}

\item{HistBreaks}{Breaks in the histogram}
}
\value{
This function will return a page of plots. The plots will show the distribution of the
   beta values with lines corresponding to the point estimate and intervals.

   A named list of the following information is also supplied:
\describe{
  \item{beta_iters}{The beta values attained from sampling}
  \item{CI_beta}{Theoretical distribution}
  \item{qbeta}{Intervals from the sampled data, compare to `CI_beta`}
  \item{iter}{Number of iterations to calculate}
  \item{alpha}{Used to create a 1-alpha interval}
  \item{failedSamples}{A list of the samples in which a singular matrix was produced}
}
}
\description{
This function will run a bootstrap analysis on a passed in model. Specifically, we
   grab a number of samples and produce \eqn{\beta} estimates of this sampled data. Occasionally it
   is not possible to create such an estimate as \eqn{X'X} is sometimes a singular matrix. The output of
   this function allows us to look at the distribution of estimates and compare them to the theoretical
   distribution.
}
\examples{
## For PRODQUAL Data
t <- MLRBootstrap(QUALITY ~ TEMP + PRESSURE + TEMP:PRESSURE + I(TEMP^2), data = PRODQUAL,
 iter = 5000, HistBreaks = 50)
t

## For Browser Data | will create singular matricies
t3 <- MLRBootstrap(TIME ~ NUMBER + BROWSER, data = BROWSER, iter = 1000)
t3
}
