% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot;skag0011Fun1.R
\name{plot.skag0011Fun1}
\alias{plot.skag0011Fun1}
\title{Plots for MLRModInfo}
\usage{
\method{plot}{skag0011Fun1}(x, y, ...)
}
\arguments{
\item{x}{x data for default plot}

\item{y}{y data for default plot}

\item{...}{more arguments}
}
\value{
4 plots in one window
}
\description{
Plots for MLRModInfo
}
\examples{
\dontrun{ plot(MLRModInfo(QUALITY ~ TEMP, PRODQUAL))}
}
