% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODQUAL.R
\docType{data}
\name{PRODQUAL}
\alias{PRODQUAL}
\title{Product Quality Data Report}
\format{
A data frame with 27 rows and 3 columns:
\describe{
  \item{TEMP}{Temperature that product was made under}
  \item{PRESSURE}{Pressure product was made under}
  \item{QUALITY}{The numeric value of the quality of the product}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
PRODQUAL
}
\description{
A set of Data relating to the quality of a product constructed at different
   pressure and temperature.
}
\keyword{datasets}
