% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyplts.R
\name{shinyplts}
\alias{shinyplts}
\title{shinyplts: Runs included shiny app}
\usage{
shinyplts()
}
\value{
Shiny Server app
}
\description{
The app allows one to select the data points on a plot by brushing (points then highlighted in orange)
   and then you can plot a new regression line for the selected data. To reset the selected points, double click. Each
   additional line will give some data output on the left about the lines. Note that selected points are still viewable
   when x and y values are changed.
}
\examples{
\dontrun{shinyplts()}
}
