% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLUDGE.R
\docType{data}
\name{SLUDGE}
\alias{SLUDGE}
\title{Sludge Data information}
\format{
A data frame with 27 rows and 4 column(s)
\describe{
  \item{CH4}{Amount of methane gas emitted by sludge (milligrams per liter)}
  \item{TIME}{Time in days the sludge was processed}
  \item{VHA}{Numeric value of yes (1) or no (0) of VHA-Added}
  \item{VHA-Added}{Yes or no depending on if VHA was added or not}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
SLUDGE
}
\description{
Data set of Browser test information.
   Data frame of 27 rows and 4 columns
}
\keyword{datasets}
