% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary;skag0011Fun1.R
\name{summary.skag0011Fun1}
\alias{summary.skag0011Fun1}
\title{summary.skag0011Fun1: addition to summary()}
\usage{
\method{summary}{skag0011Fun1}(object, ...)
}
\arguments{
\item{object}{Return form MLRModInfo()}

\item{...}{additional arguments affecting the summary produced}
}
\value{
Command line Output
}
\description{
summary.skag0011Fun1: addition to summary()
}
\examples{
\dontrun{ summary(MLRModInfo(QUALITY ~ TEMP, PRODQUAL))}
}
