% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRBootstrap.R
\name{MLRBootstrap}
\alias{MLRBootstrap}
\title{MLRBootstrap: Bootstrap analysis of MLR model}
\usage{
MLRBootstrap(model, data, iter = 5000, alpha = 0.05, HistBreaks = 30)
}
\arguments{
\item{model}{Model to do bootstrap analysis}

\item{data}{Data for bootstrap analysis}

\item{iter}{Iterations of the bootstrap}

\item{alpha}{Used to create 1-alpha CI and etc}

\item{HistBreaks}{Breaks in the histogram}
}
\value{
Fancy List
}
\description{
MLRBootstrap: Bootstrap analysis of MLR model
}
\examples{
## For PRODQUAL Data
t <- MLRBootstrap(QUALITY ~ TEMP + PRESSURE + TEMP:PRESSURE + I(TEMP^2), data = PRODQUAL,
 iter = 5000, HistBreaks = 50)
t

## For Browser Data | will create singular matricies
t3 <- MLRBootstrap(TIME ~ NUMBER + BROWSER, data = BROWSER, iter = 1000)
t3
}
