% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRModInfo.R
\name{MLRModInfo}
\alias{MLRModInfo}
\title{MLRModInfo: Gain information about a given model and find potential better models}
\usage{
MLRModInfo(model, data, alpha = 0.05)
}
\arguments{
\item{model}{Model to test}

\item{data}{data for the model to test}

\item{alpha}{Value for 1 - alpha CI}
}
\value{
Fancy List
}
\description{
MLRModInfo: Gain information about a given model and find potential better models
}
\examples{
## Example 1
ylm <- lm(QUALITY ~ TEMP + PRESSURE + I(TEMP^2) + I(PRESSURE^2) + TEMP:PRESSURE, data = PRODQUAL)
t1 <- MLRModInfo(ylm, PRODQUAL, alpha = 0.05)
summary(t1)

## Example 2 | From s20x
butterfat.fit<-lm(Butterfat~Breed+Age+Breed*Age,data=BUTTERFAT)
t2 <- MLRModInfo(butterfat.fit, BUTTERFAT)
summary(t2)
plot(t2)
}
